+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [1.0.0] - 2019-07-17
 
 * build the Web version of flip panel
+* add show days to countdown clock